requirements.txt
torch>=2.3.0
torchvision>=0.18.0
diffusers>=0.27.0
transformers>=4.40.0
accelerate>=0.30.0
peft>=0.11.0
lpips>=0.1.4
scikit-image>=0.22.0
matplotlib>=3.8.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0
pandas>=2.0.0
opencv-python>=4.8.0
einops>=0.7.0
huggingface_hub>=0.23.0
